+++ tests/control.rs
@@ -9,27 +9,32 @@ fn bel() {
     parser.process(b"\x07");
     assert_eq!(parser.screen().audible_bell_count(), 1);
     assert_eq!(parser.screen().audible_bell_count(), 1);
-    assert_eq!(parser.screen().contents_diff(&screen), b"\x07");
+    assert_eq!(parser.screen().contents_diff(&screen), b"");
+    assert_eq!(parser.screen().bells_diff(&screen), b"\x07");
 
     let screen = parser.screen().clone();
     parser.process(b"\x07");
     assert_eq!(parser.screen().audible_bell_count(), 2);
-    assert_eq!(parser.screen().contents_diff(&screen), b"\x07");
+    assert_eq!(parser.screen().contents_diff(&screen), b"");
+    assert_eq!(parser.screen().bells_diff(&screen), b"\x07");
 
     let screen = parser.screen().clone();
     parser.process(b"\x07\x07\x07");
     assert_eq!(parser.screen().audible_bell_count(), 5);
-    assert_eq!(parser.screen().contents_diff(&screen), b"\x07");
+    assert_eq!(parser.screen().contents_diff(&screen), b"");
+    assert_eq!(parser.screen().bells_diff(&screen), b"\x07");
 
     let screen = parser.screen().clone();
     parser.process(b"foo");
     assert_eq!(parser.screen().audible_bell_count(), 5);
     assert_eq!(parser.screen().contents_diff(&screen), b"foo");
+    assert_eq!(parser.screen().bells_diff(&screen), b"");
 
     let screen = parser.screen().clone();
     parser.process(b"ba\x07r");
     assert_eq!(parser.screen().audible_bell_count(), 6);
-    assert_eq!(parser.screen().contents_diff(&screen), b"bar\x07");
+    assert_eq!(parser.screen().contents_diff(&screen), b"bar");
+    assert_eq!(parser.screen().bells_diff(&screen), b"\x07");
 }
 
 #[test]